Handle case where queryRoutes takes > 0.5s + fix integ test

+++ test/integration/action/action-integration.spec.js
@@ -348,7 +348,11 @@ describe('Action Integration Tests', function() {
       const isValid = await payments1.decodeInvoice({
         invoice: store2.invoice.uri,
       });
-      expect(parseInt(store1.payment.fee), 'to be greater than or equal to', 0);
+      expect(
+        parseFloat(store1.payment.fee),
+        'to be greater than or equal to',
+        0
+      );
       expect(isValid, 'to be', true);
     });
